drop "MIT License" line from LICENSE-MIT
Since jj now uses gitoxide, I'm need to import many gitoxide crates into Google's source control system. The automation at Google doesn't like the "MIT License" line at the beginning of the LICENSE-MIT file. AFAICT, the automation is correct in thinking that the line is not supposed to be there.