feat: add `Prepare::with_shell_allow_argument_splitting()` flag.
This flag is enabled on windows by default to emulate `git` which
doesn't actually use a shell at all there.

This should help to allow windows users to use more complex programs,
which might not have been possible previously.

It also allows certain areas of `gix` to indicate that they are safe to
use with this flag, which always is the case when `git` is prefixed, like
with credential helpers.