feat: Add `Context` to provide much more context just like `git` sets it. (#1129)
`git` sets a bunch of context-carrying environment variables which `gix`
shouldn't only (and optionally) read, but also pass on to spawned processes.

With `Context` it's now possible to gather all of this information and
set it at once.

With a minimal context, one will also set the `git_dir`, particularly important
on servers, which work with many different repositories, or a clone operation
which may be in the context of one directory, but affects another.