set tested MSRV to 1.67 - the minimum required by `time`.
Thus, we set clippy to 1.70 because we can (and that was the goal
according to `helix` and is the MSRV of `home` (a cargo crate) too).

Once 1.67 isn't good enough anymore, we will update it to v1.70 so
everything matches once again. This is mainly for supporting rustsec
this one time.